+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.1.0-dev
 
 * Enhancements
+  * Add `Construct.types_of/1`
   * Support nested types in `Construct.Cast.make/3`
 
 ## v2.0.0